requirements.txt
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
MarkupSafe==2.1.1
mongoengine==0.24.2
pkgutil-resolve-name==1.3.10
pycparser==2.21
PyJWT==2.4.0
pymongo==4.2.0
pyrsistent==0.18.1
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
